+++ use_cases/question_answering/bbh/data.py
@@ -1,5 +1,7 @@
 """Prepare classes, components, functions that will prepare and post processing the data"""
 
+import re
+
 from dataclasses import dataclass, field
 import adalflow as adal
 
@@ -8,7 +10,33 @@ from adalflow.datasets.big_bench_hard import BigBenchHard
 from adalflow.utils.data import subset_dataset
 
 
-# TODO: user dont have to specify, we can auto generate a dataclass
+@dataclass
+class ObjectCountSimple(DataClass):
+    """Dataclass for string output"""
+
+    id: str = field(
+        default=None,
+        metadata={"desc": "The unique identifier of the example"},
+    )
+
+    question: str = field(
+        default=None,
+        metadata={"desc": "The question to be answered"},
+    )
+
+    answer: str = field(
+        default=None,
+        metadata={"desc": "The raw answer to the question"},  # teacher
+    )
+
+    score: float = field(
+        default=None,
+        metadata={
+            "desc": "The score of the answer, in range [0, 1]. The higher the better"
+        },
+    )
+
+
 @dataclass
 class QuestionAnswer(DataClass):
     """Dataclass for string output"""
@@ -36,25 +64,6 @@ class QuestionAnswer(DataClass):
     )  # score can be used as weight for demo, weight = score (the higher the more likely to be sampled)
 
 
-import re
-
-
-def _parse_integer_answer(answer: str):
-    """A function that parses the last integer from a string using regular expressions."""
-    try:
-        # Use regular expression to find all sequences of digits
-        numbers = re.findall(r"\d+", answer)
-        if numbers:
-            # Get the last number found
-            answer = int(numbers[-1])
-        else:
-            answer = -1
-    except ValueError:
-        answer = -1
-
-    return answer
-
-
 @adal.fun_to_component
 def parse_integer_answer(answer: str):
     """A function that parses the last integer from a string using regular expressions."""
@@ -74,7 +83,7 @@ def parse_integer_answer(answer: str):
 
 def load_datasets(max_samples: int = None):
     """Load the dataset"""
-    train_data = BigBenchHard(split="train", task_name="BBH_word_sorting")
+    train_data = BigBenchHard(split="train")
     val_data = BigBenchHard(split="val")
     test_data = BigBenchHard(split="test")
